Big one. Not sure how to summarize but here goes
Added dataloader dependency. Using it for fetching teams so that
teams with the same id are cached for the duration of the request (this
really helps when requesting the current users events and teams).
This also fixes the 403 forbidden issue I was getting for requesting
so much data at once (that is a hypothesis though).

Turning off console logging in prod.

Removing team-cowgirl.surge.sh from CORS list